fix(조우상): 데이터 타입 변환

+++ migrations/20230522081907-create-posts.js
@@ -23,7 +23,7 @@ module.exports = {
         allowNull: true
       },
       losttime: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(17, 14),
         allowNull: true
       },
       dogname: {
